change about page to a post page
