+++ apps/expo/app/index.tsx
@@ -1,22 +1,23 @@
 import React from "react";
 import { Button, Text, TextInput, TouchableOpacity, View } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { Link, Stack } from "expo-router";
+import { Stack, useRouter } from "expo-router";
 import { FlashList } from "@shopify/flash-list";
 
 import { api, type RouterOutputs } from "../src/utils/api";
 
 const PostCard: React.FC<{
   post: RouterOutputs["post"]["all"][number];
-  onPress: () => void;
   onDelete: () => void;
-}> = ({ post, onPress, onDelete }) => {
+}> = ({ post, onDelete }) => {
+  const router = useRouter();
+
   return (
     <View className="flex flex-row rounded-lg bg-white/10 p-4">
       <View className="flex-grow">
-        <TouchableOpacity onPress={onPress}>
+        <TouchableOpacity onPress={() => router.push(`/post/${post.id}`)}>
           <Text
-            className={`text-xl font-semibold text-[#cc66ff] ${
+            className={`text-xl font-semibold text-pink-400 ${
               !post.title ? "italic" : ""
             }`}
           >
@@ -65,7 +66,7 @@ const CreatePost: React.FC = () => {
         placeholder="Content"
       />
       <TouchableOpacity
-        className="rounded bg-[#cc66ff] p-2"
+        className="rounded bg-pink-400 p-2"
         onPress={() => {
           mutate({
             title,
@@ -81,38 +82,30 @@ const CreatePost: React.FC = () => {
 
 const Index = () => {
   const postQuery = api.post.all.useQuery();
-  const [showPost, setShowPost] = React.useState<string | null>(null);
 
   const deletePostMutation = api.post.delete.useMutation({
     onSettled: () => postQuery.refetch(),
   });
 
   return (
-    <SafeAreaView className="bg-[#2e026d] bg-gradient-to-b from-[#2e026d] to-[#15162c]">
+    <SafeAreaView className="bg-[#2e026d]">
       {/* Changes page title visible on the header */}
       <Stack.Screen options={{ title: "Home Page" }} />
       <View className="h-full w-full p-4">
         <Text className="mx-auto pb-2 text-5xl font-bold text-white">
-          Create <Text className="text-[#cc66ff]">T3</Text> Turbo
+          Create <Text className="text-pink-400">T3</Text> Turbo
         </Text>
 
         <Button
           onPress={() => void postQuery.refetch()}
           title="Refresh posts"
-          color={"#cc66ff"}
+          color={"#f472b6"}
         />
 
         <View className="py-2">
-          {showPost ? (
-            <Text className="text-white">
-              <Text className="font-semibold">Selected post: </Text>
-              {showPost}
-            </Text>
-          ) : (
-            <Text className="font-semibold italic text-white">
-              Press on a post
-            </Text>
-          )}
+          <Text className="font-semibold italic text-white">
+            Press on a post
+          </Text>
         </View>
 
         <FlashList
@@ -122,19 +115,12 @@ const Index = () => {
           renderItem={(p) => (
             <PostCard
               post={p.item}
-              onPress={() => setShowPost(p.item.id)}
               onDelete={() => deletePostMutation.mutate(p.item.id)}
             />
           )}
         />
 
         <CreatePost />
-
-        <View>
-          <Link href="/about">
-            <Text className="text-white">About Create T3 Turbo</Text>
-          </Link>
-        </View>
       </View>
     </SafeAreaView>
   );
