+++ apps/expo/package.json
@@ -28,8 +28,8 @@
     "react": "18.1.0",
     "react-dom": "18.1.0",
     "react-native": "0.70.5",
-    "react-native-screens": "~3.18.0",
-    "react-native-safe-area-context": "4.5.0"
+    "react-native-safe-area-context": "4.5.0",
+    "react-native-screens": "~3.18.0"
   },
   "devDependencies": {
     "@babel/core": "^7.19.3",
